Do some more code moving related to shops UI
