+++ src/Engine/mm7_data.cpp
@@ -2533,7 +2533,6 @@ std::array<unsigned int, 128 * 128 * 2> pTerrainSomeOtherData;
 unsigned int uPlayerCreationUI_SelectedCharacter;
 int currentAddressingAwardBit;
 // int dword_F8B144; // nexindex [-1] to the following
-std::array<int16_t, 6> weapons_Ypos;           // word_F8B158
 int guild_membership_approved;
 PLAYER_SKILL_MASTERY dword_F8B1B0_MasteryBeingTaught;
 int gold_transaction_amount;  // F8B1B4
